refactor: AdminUserManageServiceImpl 의 grantInitialVacations() 메서드 업데이트 (250718)
휴가 타입 하드코딩 -> 휴가 타입 조회 API로 동적 처리

+++ src/main/java/com/melly/vacationmanager/domain/vacation/balance/service/IVacationBalanceService.java
@@ -6,5 +6,5 @@ import com.melly.vacationmanager.domain.vacation.type.entity.VacationTypeEntity;
 import java.math.BigDecimal;
 
 public interface IVacationBalanceService {
-    void initializeVacationBalance(UserEntity user, VacationTypeEntity type, BigDecimal days);
+    void initializeVacationBalance(UserEntity user, VacationTypeEntity type, Integer days);
 }
